requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
psycopg2-binary>=2.9
django-environ>=0.10.0
python-dotenv>=1.0.0
django-cors-headers>=3.15.0
django-filter>=23.1
whitenoise>=6.4.0
gunicorn>=20.1.0
Pillow>=9.0.0
requests